requirements.txt
# ICP Scoring Engine - Dependencies
# ==================================

# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# HTTP Client (for external API calls)
httpx>=0.24.0
requests>=2.31.0

# LLM Providers (optional - install as needed)
openai>=1.0.0
# anthropic>=0.18.0  # Uncomment if using Anthropic

# Async Support
asyncio-throttle>=1.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: Redis for caching (production)
# redis>=4.6.0

# Optional: PostgreSQL for persistence (production)
# psycopg2-binary>=2.9.0
# sqlalchemy>=2.0.0
